requirements.txt
asgiref==3.2.10
Django==3.1.1
django-mdeditor==0.1.18
Markdown==3.2.2
Pillow==7.2.0
psycopg2==2.8.6
Pygments==2.6.1
pytz==2020.1
sqlparse==0.3.1
